Resolvido desafio de incluir Books

+++ db/migrations/V3__Create_Table_Books.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE `books` (
-  `id` INT(10) AUTO_INCREMENT PRIMARY KEY,
+  `id` bigint(20) AUTO_INCREMENT PRIMARY KEY,
   `author` longtext,
   `launch_date` datetime(6) NOT NULL,
   `price` decimal(65,2) NOT NULL,
